core: Don't fail when system partition and reserved size are defined
When BOARD_SYSTEMIMAGE_PARTITION_SIZE and
BOARD_SYSTEMIMAGE_PARTITION_RESERVED_SIZE are bothed defined, the build
will fail since it's defined in treble's common makefiles.

To avoid this, if the device has a custom system partition size
specified, set the reserved size to nothing and use the system partition
size instead. Be sure to warn the build user of this change

Change-Id: I674e7597fac34752e7231aff230b616669f181b9